1. refactored common code of data_integration_test and filter_integration_test into table_integration_Test. 2. modified filter_integration_test to use common class. 3. removed table_name argument from data_integration_test.
